Added concept images, and files for teacher creating class.
Class_Init creates the classes table that stores all of the classes
names, thresholds, and keywords.

New_Class_Form is where the teacher inserts the data for his class.

class_insert_form creates a new table for each new class. May be a bad
idea, there is mysqli and PDO.